Move cron module to independent package
Make it easier to keep track of upstream changes and bug fixes
